% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbcolors.R
\docType{data}
\name{girlboss_palettes}
\alias{girlboss_palettes}
\title{complete list of palettes}
\format{
An object of class \code{list} of length 8.
}
\usage{
girlboss_palettes
}
\description{
use \code{\link{girlboss_palette}} to construct palettes of desired length.
}
\keyword{datasets}
