% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbcolors.R
\name{girlboss_palette}
\alias{girlboss_palette}
\title{a girlboss palette generator}
\usage{
girlboss_palette(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{name of desired palette. choices are:
\code{girlboss_in_question}, \code{ur_hot_when_u_cry},  \code{cicero_space_force_mf},
\code{d1_medschool_mf}, \code{lovers_to_friends},  \code{one_night_stand_tm}, \code{thx_for_the_claws_ig},
\code{bro_waited_4_yrs_lmao}}

\item{n}{Number of colors desired. unfortunately most palettes now only
have 6 colors. But hopefully i'll add more palettes soon (tinder moment). all color
schemes are derived from my #beautiful #sexy #mind.
if omitted, uses all colors.}

\item{type}{either "continuous" or "discrete". use continuous if you want
to automatically interpolate between colours.
@importFrom graphics rgb rect par image text}
}
\value{
a vector of colours.
}
\description{
these are a handful of color palettes i made (#girl #boss #moment!!!!! #mfs #in #stem) because I can't use cute pretty colors in my thesis
unless they come from a palette. so, take that, higher education! you can't tell me what to do! even better, I named most of these
palettes after some mfs I met on tinder. i made my fun pretty palette for my thesis, but then thought that itd be even funnier to build palettes based off
of dudes i've seen this year (bc that's totally sane and professional!). i bet everyone reading through this is having so much fun seeing how
#sane i am! please admit me into your grad program! i'm so cool and sexy and smart! shoutout to Karthik (if I figure out how to link their github,
I will--i'm just really dumb ngl), whose 'wesanderson' color palette skeleton I ripped in order to build out my girboss package. xoxo thx, Karthik!
}
\examples{
girlboss_palette("girlboss_in_question")
girlboss_palette("d1_medschool_mf")
girlboss_palette("ur_hot_when_u_cry", 3)

# if you need more colors than normally found in a palette, you
# can use a continuous palette to interpolate between existing
# colors
pal <- girlboss_palette(21, name = "lovers_to_friends", type = "continuous")
image(volcano, col = pal)
}
\keyword{colors}
